refactor: add users scope to ImpactPage
