+++ src/config/routes/index.tsx
@@ -6,7 +6,7 @@ import CurrentUserProvider from "contexts/currentUserContext";
 import ConfirmEmailPage from "pages/donations/ConfirmEmailPage";
 import DonationDonePage from "pages/donations/DonationDonePage";
 import DonationInProcessPage from "pages/donations/DonationInProcessPage";
-import ImpactPage from "pages/ImpactPage";
+import ImpactPage from "pages/users/ImpactPage";
 import MainLayout from "layouts/MainLayout";
 import FundPage from "pages/promoters/FundPage";
 import SupportFundPage from "pages/promoters/SupportFundPage";
