+++ src/pages/users/ImpactPage/CardEmptySection/styles.ts
